Add `keyPrefix()` method to KeyRange, make iteratorOp() stateful
`keyPrefix(T keyPrefix)` lets you iterate forward over all entries
where the key starts with the contents of `keyPrefix`. This is useful
for compound keys.

iteratorOp() has been replaced with the class IteratorOpTester so that
the fact of having found the start key can be stored such that
future iterations don't need to test for the start key again. This
should significantly reduce the number of comparator calls for some
key range types.